requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
absl-py=0.7.1=pypi_0
astor=0.7.1=pypi_0
attrs=19.1.0=pypi_0
autograd=1.3=pypi_0
autograd-gamma=0.4.1=pypi_0
backcall=0.1.0=pypi_0
bleach=3.1.0=pypi_0
blis=0.4.1=pypi_0
boto=2.49.0=pypi_0
boto3=1.9.180=pypi_0
botocore=1.12.180=pypi_0
ca-certificates=2019.1.23=0
catalogue=1.0.0=pypi_0
catboost=0.15.2=pypi_0
certifi=2019.3.9=py36_0
chardet=3.0.4=pypi_0
click=7.1.2=pypi_0
cycler=0.10.0=pypi_0
cymem=2.0.3=pypi_0
de-core-news-sm=2.2.5=pypi_0
decorator=4.4.0=py36_1
deepctr=0.4.1=pypi_0
defusedxml=0.6.0=pypi_0
docutils=0.14=pypi_0
en-core-web-sm=2.2.0=pypi_0
entrypoints=0.3=pypi_0
fastcache=1.1.0=pypi_0
fr-core-news-sm=2.2.5=pypi_0
future=0.18.0=pypi_0
gast=0.2.2=pypi_0
gensim=3.7.3=pypi_0
grpcio=1.20.1=pypi_0
h5py=2.9.0=pypi_0
idna=2.8=pypi_0
imageio=2.6.1=pypi_0
importlib-metadata=1.6.1=pypi_0
ipykernel=5.1.0=pypi_0
ipython=7.5.0=py36h39e3cac_0
ipython-genutils=0.2.0=pypi_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.4.2=pypi_0
jax=0.1.46=pypi_0
jaxlib=0.1.31=pypi_0
jedi=0.13.3=pypi_0
jieba=0.42.1=pypi_0
jinja2=2.10.1=pypi_0
jmespath=0.9.4=pypi_0
joblib=0.14.0=pypi_0
jsonschema=3.0.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=5.2.4=pypi_0
jupyter-console=6.0.0=pypi_0
jupyter-contrib-core=0.3.3=pypi_0
jupyter-contrib-nbextensions=0.5.1=pypi_0
jupyter-core=4.4.0=pypi_0
jupyter-highlight-selected-word=0.2.0=pypi_0
jupyter-latex-envs=1.4.6=pypi_0
jupyter-nbextensions-configurator=0.4.1=pypi_0
keras=2.2.4=pypi_0
keras-applications=1.0.7=pypi_0
keras-preprocessing=1.0.9=pypi_0
kiwisolver=1.1.0=pypi_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libstdcxx-ng=8.2.0=hdf63c60_1
lifelines=0.22.8=pypi_0
lightgbm=2.2.3=pypi_0
lxml=4.3.3=pypi_0
markdown=3.1=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.1.1=pypi_0
mistune=0.8.4=pypi_0
mock=3.0.4=pypi_0
murmurhash=1.0.2=pypi_0
nbconvert=5.5.0=pypi_0
nbformat=4.4.0=pypi_0
ncurses=6.1=he6710b0_1
networkx=2.4=pypi_0
ngboost=0.1.3=pypi_0
nltk=3.5=pypi_0
notebook=5.7.8=pypi_0
numexpr=2.6.9=pypi_0
numpy=1.16.3=pypi_0
opencv-python=4.1.1.26=pypi_0
openssl=1.1.1b=h7b6447c_1
opt-einsum=3.1.0=pypi_0
pandas=0.24.2=pypi_0
pandocfilters=1.4.2=pypi_0
parso=0.4.0=py_0
pexpect=4.7.0=py36_0
pickleshare=0.7.5=pypi_0
pillow=6.0.0=pypi_0
pip=19.1=py36_0
plac=1.1.3=pypi_0
plotly=4.2.0=pypi_0
preshed=3.0.2=pypi_0
prometheus-client=0.6.0=pypi_0
prompt_toolkit=2.0.9=py36_0
protobuf=3.7.1=pypi_0
ptyprocess=0.6.0=pypi_0
pygments=2.3.1=pypi_0
pyparsing=2.4.2=pypi_0
pyrsistent=0.15.1=pypi_0
python=3.6.8=h0371630_0
python-dateutil=2.8.0=pypi_0
python-graphviz=0.13=pypi_0
pytz=2019.1=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.1=pypi_0
pyzmq=18.0.1=pypi_0
qtconsole=4.4.4=pypi_0
readline=7.0=h7b6447c_5
regex=2020.6.8=pypi_0
requests=2.22.0=pypi_0
retrying=1.3.3=pypi_0
s3transfer=0.2.1=pypi_0
scikit-image=0.16.2=pypi_0
scikit-learn=0.21.3=pypi_0
scipy=1.2.1=pypi_0
send2trash=1.5.0=pypi_0
setuptools=41.0.1=py36_0
six=1.12.0=pypi_0
smart-open=1.8.4=pypi_0
some-package=0.1=pypi_0
spacy=2.3.0=pypi_0
sqlite=3.28.0=h7b6447c_0
srsly=1.0.2=pypi_0
stanza=1.0.1=pypi_0
tables=3.5.2=pypi_0
tensorboard=1.12.2=pypi_0
tensorflow-estimator=1.13.0=pypi_0
tensorflow-gpu=1.12.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.2=pypi_0
testpath=0.4.2=pypi_0
thinc=7.4.1=pypi_0
tk=8.6.8=hbc83047_0
torch=1.6.0=pypi_0
torchvision=0.7.0=pypi_0
tornado=6.0.2=pypi_0
tqdm=4.46.1=pypi_0
traitlets=4.3.2=py36_0
urllib3=1.25.3=pypi_0
wasabi=0.6.0=pypi_0
wcwidth=0.1.7=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=0.15.2=pypi_0
wheel=0.33.1=py36_0
widgetsnbextension=3.4.2=pypi_0
xgboost=0.82=pypi_0
xx-ent-wiki-sm=2.2.0=pypi_0
xz=5.2.4=h14c3975_4
zh-core-web-sm=2.3.1=pypi_0
zipp=3.1.0=pypi_0
zlib=1.2.11=h7b6447c_3
